Use ES6 Module & luxon for navTime
